+++ examples/inference.py
@@ -1,3 +1,9 @@
+"""
+How to use the inference API.
+
+This example downloads an image from an URL and uses a pretrained model on Pets.
+"""
+
 from mantisshrimp.imports import *
 from mantisshrimp import *
 from mantisshrimp.models.rcnn import faster_rcnn
@@ -18,7 +24,7 @@ img = open_img(IMG_PATH)
 show_img(img)
 
 # The model was trained with normalized images, it's necessary to do the same in inference
-tfms = AlbuTransform([A.Normalize()])
+tfms = AlbumentationTransforms([A.Normalize()])
 
 # Whenever you have images in memory (numpy arrays) you can use `Dataset.from_images`
 infer_ds = Dataset.from_images([img], tfms)