+++ mantisshrimp/transforms/__init__.py
@@ -1,4 +1,4 @@
 from mantisshrimp.transforms.transform import *
-from mantisshrimp.transforms.albu_transform import *
+from mantisshrimp.transforms.albumentation_transforms import *
 
 import mantisshrimp.transforms.batch_tfms