+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Renames `AlbuTransform` to `AlbumentationTransforms`
 - All `build_batch` method now returns `batch, samples`, the batch is always a tuple of inputs to the model
 
 ## [0.0.0-pre-release]
@@ -44,4 +45,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
 ## Links  
-[Unreleased]: https://github.com/lgvaz/mantisshrimp/tree/master
+[Unreleased]: https://github.com/airctic/mantisshrimp/tree/master